requirements.txt
appdirs==1.4.4
asgiref==3.2.10
black==20.8b1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dataclasses==0.7
ddf==0.0.0
defusedxml==0.6.0
Django==3.1.1
django-allauth==0.42.0
django-cors-headers==3.5.0
django-dynamic-fixture==3.1.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.11.1
idna==2.10
mypy-extensions==0.4.3
oauthlib==3.1.0
pathspec==0.8.0
python3-openid==3.2.0
pytz==2020.1
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
